Move example to root
